pd: use tower-actor for Consensus, Mempool
Now that we have a basic implementation of `tower-actor`, this PR replaces our
existing ad-hoc `Service` impls with it.
